Should have fixed bug that caused a crash after UIN32_T_MAX microseconds (~1 hour 11 minutes)

+++ src/PIOMatrixOutput/pio_matrix_output.h
@@ -28,7 +28,7 @@ class MatrixOutput {
 
     dma_channel_config dmac;
 
-    uint32_t lastSendData;
+    uint64_t lastSendData;
 
 
     uint32_t frameBuffer[MAX_FRAMES_IN_BUFFER][MAX_SUBFRAMES][MATRIX_SIZE * 3 / 4];
